Move stale data and interpolation into the quality module
Create a new gaps submodule in quality for functions that identify
gaps/missing data.

Update README to describe the quality.gaps module.
