+++ pvanalytics/tests/quality/test_gaps.py
@@ -1,7 +1,8 @@
+"""Tests for gaps quality control functions."""
 import pytest
 import pandas as pd
 from pandas.util.testing import assert_series_equal
-from pvanalytics import filtering
+from pvanalytics.quality import gaps
 
 
 @pytest.fixture
@@ -20,13 +21,13 @@ def data_with_negatives():
 
 def test_detect_stale_values(stale_data):
     """detect_stale_values properly identifies stuck values."""
-    res1 = filtering.detect_stale_values(stale_data)
-    res2 = filtering.detect_stale_values(stale_data, rtol=1e-8, window=2)
-    res3 = filtering.detect_stale_values(stale_data, window=7)
-    res4 = filtering.detect_stale_values(stale_data, window=8)
-    res5 = filtering.detect_stale_values(stale_data, rtol=1e-8, window=4)
-    res6 = filtering.detect_stale_values(stale_data[1:])
-    res7 = filtering.detect_stale_values(stale_data[1:8])
+    res1 = gaps.detect_stale_values(stale_data)
+    res2 = gaps.detect_stale_values(stale_data, rtol=1e-8, window=2)
+    res3 = gaps.detect_stale_values(stale_data, window=7)
+    res4 = gaps.detect_stale_values(stale_data, window=8)
+    res5 = gaps.detect_stale_values(stale_data, rtol=1e-8, window=4)
+    res6 = gaps.detect_stale_values(stale_data[1:])
+    res7 = gaps.detect_stale_values(stale_data[1:8])
     assert_series_equal(res1, pd.Series([False, False, False, True, True, True,
                                          True, True, False, False]))
     assert_series_equal(res2, pd.Series([False, False, True, True, True, False,
@@ -43,25 +44,27 @@ def test_detect_stale_values(stale_data):
                                         data=[False, False, True, True, True,
                                               True, True]))
 
+
 def test_detect_stale_values_handles_negatives(data_with_negatives):
     """detect_stale_values works with negative values."""
-    res = filtering.detect_stale_values(data_with_negatives)
+    res = gaps.detect_stale_values(data_with_negatives)
     assert_series_equal(res, pd.Series([False, False, True, True, False, False,
                                         False]))
-    res = filtering.detect_stale_values(data_with_negatives, atol=1e-3)
+    res = gaps.detect_stale_values(data_with_negatives, atol=1e-3)
     assert_series_equal(res, pd.Series([False, False, True, True, True, True,
                                         True]))
-    res = filtering.detect_stale_values(data_with_negatives, atol=1e-5)
+    res = gaps.detect_stale_values(data_with_negatives, atol=1e-5)
     assert_series_equal(res, pd.Series([False, False, True, True, True, False,
                                         False]))
-    res = filtering.detect_stale_values(data_with_negatives, atol=2e-5)
+    res = gaps.detect_stale_values(data_with_negatives, atol=2e-5)
     assert_series_equal(res, pd.Series([False, False, True, True, True, True,
                                         True]))
+
 
 def test_detect_stale_values_raises_error(stale_data):
     """detect_stale_values raises a ValueError for 'window' < 2."""
     with pytest.raises(ValueError):
-        filtering.detect_stale_values(stale_data, window=1)
+        gaps.detect_stale_values(stale_data, window=1)
 
 
 @pytest.fixture
@@ -74,37 +77,39 @@ def interpolated_data():
 
 def test_detect_interpolation(interpolated_data):
     """Interpolation is detected correclty."""
-    res1 = filtering.detect_interpolation(interpolated_data)
+    res1 = gaps.detect_interpolation(interpolated_data)
     assert_series_equal(res1, pd.Series([False, False, False, False, False,
                                          False, False, True, False, False,
                                          False, False, False, True, True, True,
                                          False]))
-    res2 = filtering.detect_interpolation(interpolated_data, rtol=1e-2)
+    res2 = gaps.detect_interpolation(interpolated_data, rtol=1e-2)
     assert_series_equal(res2, pd.Series([False, False, True, True, True,
                                          False, False, True, False, False,
                                          False, False, False, True, True, True,
                                          False]))
-    res3 = filtering.detect_interpolation(interpolated_data, window=5)
+    res3 = gaps.detect_interpolation(interpolated_data, window=5)
     assert_series_equal(res3, pd.Series([False, False, False, False, False,
                                          False, False, False, False, False,
                                          False, False, False, False, False,
                                          True, False]))
-    res4 = filtering.detect_interpolation(interpolated_data, atol=1e-2)
+    res4 = gaps.detect_interpolation(interpolated_data, atol=1e-2)
     assert_series_equal(res4, pd.Series([False, False, True, True, True,
                                          True, True, True, False, False,
                                          False, False, False, True, True, True,
                                          False]))
 
+
 def test_detect_interpolation_handles_negatives(data_with_negatives):
     """Interpolation is detected correctly when data contains negatives."""
-    res = filtering.detect_interpolation(data_with_negatives, atol=1e-5)
+    res = gaps.detect_interpolation(data_with_negatives, atol=1e-5)
     assert_series_equal(res, pd.Series([False, False, True, True, True, True,
                                         False]))
-    res = filtering.detect_stale_values(data_with_negatives, atol=1e-4)
+    res = gaps.detect_stale_values(data_with_negatives, atol=1e-4)
     assert_series_equal(res, pd.Series([False, False, True, True, True, True,
                                         True]))
+
 
 def test_detect_interpolation_raises_error(interpolated_data):
     """detect_interpolation raises a ValueError for 'window' < 3."""
     with pytest.raises(ValueError):
-        filtering.detect_interpolation(interpolated_data, window=2)
+        gaps.detect_interpolation(interpolated_data, window=2)
